+++ src/DataSources/Pdf/Chunker/PdfChunk.cs
@@ -3,9 +3,26 @@
 /// <summary>
 /// Represent a chunk of a PDF
 /// </summary>
-/// <param name="Filename">Filename of the PDF</param>
-/// <param name="Folder">Folder of the PDF</param>
-/// <param name="Page">Page of the PDF</param>
-/// <param name="TotalPages">Total Pages in the PDF</param>
-/// <param name="Text">Raw Text of the PDF</param>
-public record PdfChunk(string Filename, string Folder, int Page, int TotalPages, string Text);
+/// <param name="name">Name of the PDF</param>
+/// <param name="page">Page of the PDF</param>
+/// <param name="totalPages">Total Pages in the PDF</param>
+/// <param name="text">Raw Text of the PDF</param>
+public class PdfChunk(string name, int page, int totalPages, string text)
+{
+    /// <summary>Name of the PDF</summary>
+    public string Name { get; } = name;
+
+    /// <summary>Page of the PDF</summary>
+    public int Page { get; } = page;
+
+    /// <summary>Total Pages in the PDF</summary>
+    public int TotalPages { get; } = totalPages;
+
+    /// <summary>Raw Text of the PDF</summary>
+    public string Text { get; } = text;
+
+    /// <summary>
+    /// SourcePath ot the chunk
+    /// </summary>
+    public string SourcePath { get; set; } = "";
+}